requirements.txt
pystray~=0.19.5
pillow~=11.1.0
pynput~=1.8.0
opencv-python~=4.11.0.86
numpy~=2.2.3
comtypes~=1.4.10
pycaw==20240210
PyYAML~=6.0.2
